requirements.txt
pandas
openpyxl
tabulate
bs4
ollama
python-dotenv
azure-identity
azure-search-documents==11.4b9
streamlit